hex-fa: factor out ukernels into separate headers and unify the rest
