+++ ggml/src/ggml-hexagon/htp/CMakeLists.txt
@@ -21,7 +21,6 @@ add_library(${HTP_LIB} SHARED
     hex-dma.c
     hmx-queue.c
     flash-attn-ops.c
-    hmx-flash-attn-ops.c
     matmul-ops.c
     binary-ops.c
     unary-ops.c
